backend/requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
fastapi==0.119.0
filelock==3.20.0
filetype==1.2.0
# Install this before langchain-google-genai if resolver conflicts occur.
google-ai-generativelanguage
google-api-core>=2,<3
google-api-python-client>=2,<3
google-auth>=2,<3
google-auth-httplib2>=0.2
# Install this before langchain-google-genai if resolver conflicts occur.
google-generativeai
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-core
# Might require manual installation to avoid dependency resolver conflicts.
langchain-google-genai
langchain-qdrant
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
langchain-text-splitters
looseversion==1.3.0
lxml==6.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==12.0.0
pip==25.2
portalocker==3.2.0
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311; sys_platform == "win32"
pyxnat==1.6.3
PyYAML==6.0.3
qdrant-client==1.15.1
rdflib==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.16.2
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tavily-python
tenacity==9.1.2
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
